Extract the RegistrationCostUpdater
