+++ lib/tasks/scheduler.rake
@@ -1,7 +1,7 @@
 desc "This task is called by the Heroku scheduler add-on"
 task update_registration_period: :environment do
   puts "Updating RegistrationPeriod:current_period ..."
-  RegistrationCost.update_registration_periods
+  RegistrationCostUpdater.update_registration_periods
   puts "done."
 end
